+++ src/Command.php
@@ -3,9 +3,14 @@ declare(strict_types = 1);
 
 namespace Innmind\CLI;
 
+use Innmind\Immutable\Attempt;
+
 interface Command
 {
-    public function __invoke(Console $console): Console;
+    /**
+     * @return Attempt<Console>
+     */
+    public function __invoke(Console $console): Attempt;
 
     /**
      * @psalm-mutation-free