+++ src/Environment.php
@@ -9,6 +9,7 @@ use Innmind\Immutable\{
     Map,
     Sequence,
     Str,
+    Attempt,
     Maybe,
 };
 
@@ -25,11 +26,19 @@ interface Environment
     /**
      * @param ?positive-int $length
      *
-     * @return array{Maybe<Str>, self}
+     * @return array{Attempt<Str>, self}
      */
     public function read(?int $length = null): array;
-    public function output(Str $data): self;
-    public function error(Str $data): self;
+
+    /**
+     * @return Attempt<self>
+     */
+    public function output(Str $data): Attempt;
+
+    /**
+     * @return Attempt<self>
+     */
+    public function error(Str $data): Attempt;
 
     /**
      * @return Sequence<string>