+++ tests/CommandsTest.php
@@ -9,6 +9,7 @@ use Innmind\CLI\{
     Environment,
     Console,
 };
+use Innmind\Immutable\Attempt;
 use Innmind\BlackBox\PHPUnit\Framework\TestCase;
 
 class CommandsTest extends TestCase
@@ -16,7 +17,7 @@ class CommandsTest extends TestCase
     public function testRunSingleCommand()
     {
         $run = Commands::of(new class implements Command {
-            public function __invoke(Console $console): Console
+            public function __invoke(Console $console): Attempt
             {
                 if (
                     !$console->arguments()->contains('container') ||
@@ -25,10 +26,10 @@ class CommandsTest extends TestCase
                     $console->arguments()->get('output') !== 'bar' ||
                     !$console->options()->contains('foo')
                 ) {
-                    throw new \Exception;
+                    return Attempt::error(new \Exception);
                 }
 
-                return $console->exit(42);
+                return Attempt::result($console->exit(42));
             }
 
             public function usage(): string
@@ -44,7 +45,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $this->assertSame(42, $run($env)->exitCode()->match(
+        $this->assertSame(42, $run($env)->unwrap()->exitCode()->match(
             static fn($code) => $code->toInt(),
             static fn() => null,
         ));
@@ -54,9 +55,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -65,7 +66,7 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
                     if (
                         !$console->arguments()->contains('container') ||
@@ -74,10 +75,10 @@ class CommandsTest extends TestCase
                         $console->arguments()->get('output') !== 'bar' ||
                         !$console->options()->contains('foo')
                     ) {
-                        throw new \Exception;
+                        return Attempt::error(new \Exception);
                     }
 
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -94,7 +95,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $this->assertSame(24, $run($env)->exitCode()->match(
+        $this->assertSame(24, $run($env)->unwrap()->exitCode()->match(
             static fn($code) => $code->toInt(),
             static fn() => null,
         ));
@@ -104,9 +105,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -115,9 +116,9 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -134,7 +135,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $this->assertSame(42, $run($env)->exitCode()->match(
+        $this->assertSame(42, $run($env)->unwrap()->exitCode()->match(
             static fn($code) => $code->toInt(),
             static fn() => null,
         ));
@@ -144,9 +145,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -155,9 +156,9 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -174,7 +175,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $this->assertSame(24, $run($env)->exitCode()->match(
+        $this->assertSame(24, $run($env)->unwrap()->exitCode()->match(
             static fn($code) => $code->toInt(),
             static fn() => null,
         ));
@@ -184,9 +185,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -195,9 +196,9 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -214,7 +215,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $this->assertSame(42, $run($env)->exitCode()->match(
+        $this->assertSame(42, $run($env)->unwrap()->exitCode()->match(
             static fn($code) => $code->toInt(),
             static fn() => null,
         ));
@@ -224,9 +225,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -235,9 +236,9 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -254,7 +255,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertSame(64, $env->exitCode()->match(
             static fn($code) => $code->toInt(),
@@ -273,9 +274,9 @@ class CommandsTest extends TestCase
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -284,9 +285,9 @@ class CommandsTest extends TestCase
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -303,7 +304,7 @@ class CommandsTest extends TestCase
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertSame(64, $env->exitCode()->match(
             static fn($code) => $code->toInt(),
@@ -321,9 +322,9 @@ class CommandsTest extends TestCase
     public function testExitWhenCommandMisused()
     {
         $run = Commands::of(new class implements Command {
-            public function __invoke(Console $console): Console
+            public function __invoke(Console $console): Attempt
             {
-                return $console->exit(42);
+                return Attempt::result($console->exit(42));
             }
 
             public function usage(): string
@@ -345,7 +346,7 @@ USAGE;
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertSame(64, $env->exitCode()->match(
             static fn($code) => $code->toInt(),
@@ -360,9 +361,9 @@ USAGE;
     public function testEnvNotTemperedWhenCommandThrows()
     {
         $run = Commands::of(new class implements Command {
-            public function __invoke(Console $console): Console
+            public function __invoke(Console $console): Attempt
             {
-                throw new \Exception;
+                return Attempt::error(new \Exception);
             }
 
             public function usage(): string
@@ -380,15 +381,15 @@ USAGE;
 
         $this->expectException(\Exception::class);
 
-        $run($env);
+        $run($env)->unwrap();
     }
 
     public function testDisplayUsageWhenHelpOptionFound()
     {
         $run = Commands::of(new class implements Command {
-            public function __invoke(Console $console): Console
+            public function __invoke(Console $console): Attempt
             {
-                return $console->exit(42);
+                return Attempt::result($console->exit(42));
             }
 
             public function usage(): string
@@ -410,7 +411,7 @@ USAGE;
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertNull($env->exitCode()->match(
             static fn($code) => $code->toInt(),
@@ -426,9 +427,9 @@ USAGE;
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -437,9 +438,9 @@ USAGE;
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -456,7 +457,7 @@ USAGE;
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertNull($env->exitCode()->match(
             static fn($code) => $code->toInt(),
@@ -475,9 +476,9 @@ USAGE;
     {
         $run = Commands::of(
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(42);
+                    return Attempt::result($console->exit(42));
                 }
 
                 public function usage(): string
@@ -486,9 +487,9 @@ USAGE;
                 }
             },
             new class implements Command {
-                public function __invoke(Console $console): Console
+                public function __invoke(Console $console): Attempt
                 {
-                    return $console->exit(24);
+                    return Attempt::result($console->exit(24));
                 }
 
                 public function usage(): string
@@ -505,7 +506,7 @@ USAGE;
             '/',
         );
 
-        $env = $run($env);
+        $env = $run($env)->unwrap();
 
         $this->assertSame(64, $env->exitCode()->match(
             static fn($code) => $code->toInt(),