+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 
 - Requires `innmind/immutable:~5.16`
 - Requires `innmind/operating-system:~6.0`
+- `Innmind\Cli\Environment::read()` now return the read data in an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Console::read()` now return the read data in an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Console::output()` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Console::error()` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Environment::output()` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Environment::error()` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\Question` now return the read data in an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\ChoiceQuestion` now return the read data in an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\Question` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Question\ChoiceQuestion` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Commands` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Command` now return an `Innmind\Immutable\Attempt`
+- `Innmind\Cli\Main::main()` now return an `Innmind\Immutable\Attempt`
 
 ### Fixed